requirements.txt
django
goslate
polib
six
google-cloud-translate==2.0
google-oauth
